refactor: Context層をProvider層に分離（ISP準拠）
DatabaseContext(1330行)を10個の単責任Providerに分解:
Connection, Query, AsyncQuery, Schema, Transaction,Export, Search, Utility, Settings, IO

- IConnectionProvider参照注入でProvider間依存を解決
- interfaces/ を providers/ サブディレクトリに再編成
- JsonUtils に buildArray/buildRowArray テンプレート追加
