+++ backend/contexts/system_context.cpp
@@ -1,39 +1,60 @@
 #include "system_context.h"
 
-#include "database_context.h"
-#include "export_context.h"
-#include "io_context.h"
-#include "settings_context.h"
-#include "utility_context.h"
+#include "../providers/async_query_provider.h"
+#include "../providers/connection_provider.h"
+#include "../providers/export_provider.h"
+#include "../providers/io_provider.h"
+#include "../providers/query_provider.h"
+#include "../providers/schema_provider.h"
+#include "../providers/search_provider.h"
+#include "../providers/settings_provider.h"
+#include "../providers/transaction_provider.h"
+#include "../providers/utility_provider.h"
 
 namespace velocitydb {
 
 SystemContext::SystemContext()
-    : m_database(std::make_unique<DatabaseContext>())
-    , m_export(std::make_unique<ExportContext>())
-    , m_settings(std::make_unique<SettingsContext>())
-    , m_utility(std::make_unique<UtilityContext>())
-    , m_io(std::make_unique<IOContext>()) {}
+    : m_connections(std::make_unique<ConnectionProvider>())
+    , m_queries(std::make_unique<QueryProvider>(*m_connections))
+    , m_asyncQueries(std::make_unique<AsyncQueryProvider>(*m_connections))
+    , m_schema(std::make_unique<SchemaProvider>(*m_connections))
+    , m_transactions(std::make_unique<TransactionProvider>(*m_connections))
+    , m_exports(std::make_unique<ExportProvider>(*m_connections))
+    , m_search(std::make_unique<SearchProvider>(*m_connections))
+    , m_utility(std::make_unique<UtilityProvider>())
+    , m_settings(std::make_unique<SettingsProvider>())
+    , m_io(std::make_unique<IOProvider>()) {}
 
 SystemContext::~SystemContext() = default;
 
-IDatabaseContext& SystemContext::database() noexcept {
-    return *m_database;
+IConnectionProvider& SystemContext::connections() noexcept {
+    return *m_connections;
 }
-
-IExportContext& SystemContext::export_ctx() noexcept {
-    return *m_export;
+IQueryProvider& SystemContext::queries() noexcept {
+    return *m_queries;
 }
-
-ISettingsContext& SystemContext::settings() noexcept {
-    return *m_settings;
+IAsyncQueryProvider& SystemContext::async_queries() noexcept {
+    return *m_asyncQueries;
 }
-
-IUtilityContext& SystemContext::utility() noexcept {
+ISchemaProvider& SystemContext::schema() noexcept {
+    return *m_schema;
+}
+ITransactionProvider& SystemContext::transactions() noexcept {
+    return *m_transactions;
+}
+IExportProvider& SystemContext::exports() noexcept {
+    return *m_exports;
+}
+ISearchProvider& SystemContext::search() noexcept {
+    return *m_search;
+}
+IUtilityProvider& SystemContext::utility() noexcept {
     return *m_utility;
 }
-
-IIOContext& SystemContext::io() noexcept {
+ISettingsProvider& SystemContext::settings() noexcept {
+    return *m_settings;
+}
+IIOProvider& SystemContext::io() noexcept {
     return *m_io;
 }
