+++ backend/interfaces/system_context.h
@@ -2,22 +2,32 @@
 
 namespace velocitydb {
 
-class IDatabaseContext;
-class IExportContext;
-class ISettingsContext;
-class IUtilityContext;
-class IIOContext;
+class IConnectionProvider;
+class IQueryProvider;
+class IAsyncQueryProvider;
+class ISchemaProvider;
+class ITransactionProvider;
+class IExportProvider;
+class ISearchProvider;
+class IUtilityProvider;
+class ISettingsProvider;
+class IIOProvider;
 
-/// Top-level interface aggregating all context interfaces
+/// Top-level interface aggregating all provider interfaces
 class ISystemContext {
 public:
     virtual ~ISystemContext() = default;
 
-    [[nodiscard]] virtual IDatabaseContext& database() noexcept = 0;
-    [[nodiscard]] virtual IExportContext& export_ctx() noexcept = 0;
-    [[nodiscard]] virtual ISettingsContext& settings() noexcept = 0;
-    [[nodiscard]] virtual IUtilityContext& utility() noexcept = 0;
-    [[nodiscard]] virtual IIOContext& io() noexcept = 0;
+    [[nodiscard]] virtual IConnectionProvider& connections() noexcept = 0;
+    [[nodiscard]] virtual IQueryProvider& queries() noexcept = 0;
+    [[nodiscard]] virtual IAsyncQueryProvider& async_queries() noexcept = 0;
+    [[nodiscard]] virtual ISchemaProvider& schema() noexcept = 0;
+    [[nodiscard]] virtual ITransactionProvider& transactions() noexcept = 0;
+    [[nodiscard]] virtual IExportProvider& exports() noexcept = 0;
+    [[nodiscard]] virtual ISearchProvider& search() noexcept = 0;
+    [[nodiscard]] virtual IUtilityProvider& utility() noexcept = 0;
+    [[nodiscard]] virtual ISettingsProvider& settings() noexcept = 0;
+    [[nodiscard]] virtual IIOProvider& io() noexcept = 0;
 };
 
 }  // namespace velocitydb
